requirements.txt
pandas 
numpy
seaborn
matplotlib
scikit-learn
statsmodels
scipy
plotly
Pyyaml
xgboost
lightgbm
catboost